Adicionando tratamento de erro quando não é encontrado nada no localStorage

+++ gamagit/src/Pages/Repositories/index.js
@@ -1,15 +1,20 @@
 import React, { useEffect, useState } from 'react';
+import { useHistory } from 'react-router-dom';
 import { Title, Container, List, ListItem, LinkHome } from './styled';
 
 export default function Repositories() {
-
+    const history = useHistory();
     const [repositories, setRepositories] = useState([]);
     useEffect(() => {
         let [repositoriesName, setRepositories] = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
 
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if (repositoriesName !== null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            history.push('/')
+        }
 
     }, [])
     return (
